better file for methods
